added 1.5.5-x86_64-ubuntu-19.04

+++ include/vnx/RecordReader.h
@@ -101,7 +101,7 @@ private:
 	std::shared_ptr<RecordHeader> header;
 	std::shared_ptr<RecordIndex> curr_block;
 	std::shared_ptr<Sample> curr_sample;
-	int64_t index_pos = -1;
+	int32_t index_pos = -1;
 	
 	bool is_begin_of_file = false;
 	bool is_end_of_file = false;
